LFS-938: Move the BioPortalApiKey serverside code into a separate API component
